### 2022-11-03 #115 @whedon
[PRE REVIEW]: Distributed Parallelization of xPU Stencil Computations in Julia
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** <a href="https://github.com/omlins/ImplicitGlobalGrid.jl" target ="_blank">https://github.com/omlins/ImplicitGlobalGrid.jl</a>
**Version:**
