+++ paper/paper.yml
@@ -9,13 +9,13 @@ keywords:
   - Staggered Grid
 authors:
   - name: Samuel Omlin
-    orcid: 0000-0000-0000-000X
+    orcid: 0000-0001-8969-5619
     affiliation: 1
   - name: Ludovic R\"ass
     orcid: 0000-0002-1136-899X
     affiliation: 2, 3
   - name: Ivan Utkin
-    orcid: 0000-0000-0000-000X
+    orcid: 0000-0002-5233-6783
     affiliation: 2, 3
 affiliations:
   - name: Swiss National Supercomputing Centre (CSCS), ETH Zurich, Lugano, Switzerland
