### 2023-03-05 #115 @smartalecH
@omlins the manuscript mentions all experiments benchmark _weak scaling_ (where the global size of the problem increases with resource count such that the burden on each individual resource remains constant) but the linked repo seems to indicate that the global problem size stays _fixed_ as a function of GPU count. Namely, the `jl` script and the data itself seem to indicate that a global problem size of `512x512x512` is always used (independent of GPU/process count, such that each local domain shrinks with larger resource counts).

Am I misinterpreting something? Perhaps with how the local vs global grid is defined?

I also only see the data/codes for the diffusion experimental results. Perhaps the other experimental results are stored elsewhere?

(Sorry to hijack the thread a bit... but this seemed the most relevant place to post...)

Thanks!
