### 2023-03-05 #115 @smartalecH
> Namely, the jl script and the data itself seem to indicate that a global problem size of 512x512x512 is always used (independent of GPU/process count, such that each local domain shrinks with larger resource counts).

Ah it seems this is where the "implicit" part of the global grid comes into play. In other words, the user always specifies the _local_ grid, and those parameters are used to implicitly generate a global grid where each process uses identical parameters.

This is a bit different paradigm than what I'm used to. Usually we specify the global grid first and choose a distributed topology based on available resources etc (which also allows one to perform load balancing, optimize for chunks that minimize communication overhead, etc). But there's a lot of value to the simplicity offered here. Especially since we often run distributed simulations on identical hardware and want to maximize each individual node.
