### 2023-03-28 #115 @omlins
> Ah it seems this is where the "implicit" part of the global grid comes into play. In other words, the user always specifies the local grid, and those parameters are used to implicitly generate a global grid where each process uses identical parameters.

I'm glad you figured it out yourself! :+1:
