### 2023-10-25 #115 @fcdimitr
Hi @omlins. I am giving a quick update from my side.

I found a first reviewer. I'll try to find a second one ASAP.

@georgebisbas, thank you for volunteering as a reviewer. You can find more details on the JuliaCon review process [here](https://juliacon.github.io/proceedings-guide/reviewer/). Note this submission is still in pre-review, and the review will start once I find a second reviewer. Meanwhile, feel free to ask me any questions you have.
