### 2023-11-06 #137 @whedon
[REVIEW]: Distributed Parallelization of xPU Stencil Computations in Julia
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** https://github.com/omlins/ImplicitGlobalGrid.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] #167

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
