### 2023-11-20 #137 @mloubout
# Review

The abstract is written clearly and convey the message efficiently. The motivation and design principles are laid out clearly. The results show good performance and validate the claims of the author.

Some minor comments:
- "We observe a parallel efficiency of 93% on 2197 GPUs." this is an odd number of GPUs to use for scaling, why not use 1024 similar to the parallel efficiency
- Fig 3 could use the performance comparison since the author admit that the CUDA implementation outperforms the Juia implementation and this figure alone is slightly miss-leading.
- Staggered grid can lead to asymmetrical halos (potentially send only/recv only), a small explanation on how it is handled would be useful
- How does it compare to similar DSLs for distributed stencil/pde solving such as [Firedrake](https://www.firedrakeproject.org/index.html), [Fenics](https://fenicsproject.org/), [Devito](https://www.devitoproject.org/)

very minor:
- The overlap with https://github.com/JuliaCon/proceedings-review/issues/138 is non-negligeable, I wonder if those two should be combined.

With these addressed, this should be accepted for publication
