### 2023-11-25 #137 @georgebisbas
It seems like the repository: https://github.com/omlins/ImplicitGlobalGrid.jl
does not have an active "Issues" Github tab. Is that expected?
