### 2023-12-05 #137 @georgebisbas
> > It seems like the repository: https://github.com/omlins/ImplicitGlobalGrid.jl does not have an active "Issues" Github tab. Is that expected?
>
> Yes this is expected, because it is a fork for the single purpose of the publication of this paper. The original repository is the following: https://github.com/eth-cscs/ImplicitGlobalGrid.jl

Thanks @omlins. Where should we open an Issue for the paper according to the instructions? At the original repo?
