### 2023-12-05 #137 @georgebisbas
> I do not think that for this configuration it is specified in the instructions. Thus, I would suggest to open issues concerning the manuscript on the fork (https://github.com/omlins/ImplicitGlobalGrid.jl) and potential issues concerning the package on the original repository (https://github.com/eth-cscs/ImplicitGlobalGrid.jl).

Now I can see a tab of Issues:
https://github.com/omlins/ImplicitGlobalGrid.jl/issues
which was not there before. Thanks
