### 2024-01-15 #137 @georgebisbas
ImplcitGlobalGrid.jl offers an approach to automate distributed memory parallelism for stencil computations in Julia, leveraging a high-level symbolic interface and presenting highly competitive performance.

This work has a high overlap with  https://github.com/JuliaCon/proceedings-review/issues/138

As discussed in ParallelStencil.jI, some changes should be applied to publish those as separate works. I understand that this work extends the optimization suite available in ParallelStencl.jl by allowing the use of automated Distributed Memory Parallelism. Since both packages will be published independently, mentioning that DMP is built upon the ParallelStencil.jl would not harm.

I suggest avoiding using the term xPU in the title. In the recent rise of “exotic” accelerators like Google’s TPU, Graphcore’s IPU, or expected QPUs, it probably would not harm just to use:

"Distributed Parallelization of Stencil Computations in Julia"

This is clarified at the end of the Introduction. However, I still advocate avoiding this term. It may also not help when someone will be looking for your work in the future.

The package's functionality is confirmed, and the installation, testing, and use documentation are present and well-defined. The paper is well-written, clearly presenting the motivation and example usage of the package. It may be good to add explicitly that users get automated DMP by only slightly editing the source code of a solver. I.e., the text can value the symbolic interface a bit more.

I am interested in how communication computation overlap is performed.
Would adding some more details about this in Section 2 be easy?

The diffusion example presented is well-written. Minor comments (not aiming to be addressed, but possibly interesting points for future work and further discussion):
- Should Figure 1 be presented as a Listing 1?

Performance evaluation:
- Why are you using 2197 GPUs?
- Why is CPU scaling not presented since the paper claims xPUs?
- Why is strong scaling not presented in this work?
- How good is single-node performance?

Good to see that you are using a non-standard stencil kernel in Figure 3. Please add more info on this stencil kernel's computation and memory requirements.

I also second the comments posted by @mloubout.

Assuming the above concerns are addressed, this should be accepted for publication.
