### 2024-02-27 #137 @georgebisbas
> The single-node performance is very high thanks to the implementation using ParallelStencil.jl.

Are there any performance numbers, like Gpts/s or GCells/s or any roofline model, we could look at?
