+++ paper/ref.bib
@@ -14,18 +14,14 @@
   booktitle={SC20: International Conference for High Performance Computing, Networking, Storage and Analysis},   
   title={Fast Stencil-Code Computation on a Wafer-Scale Processor},   
   year={2020},  
-  volume={},  
-  number={},  
   pages={1-14},  
   doi={10.1109/SC41405.2020.00062}
 }
 
 @article{parallelstencil2022,
   author={Omlin, S. and R{\"{a}}ss, L.},
-  doi = {???},
   journal = {Proc. JuliaCon Conf.},
   title = {{H}igh-performance x{PU} {S}tencil {C}omputations in {J}ulia},
-  volume = {?},
   pages = {2},
   year = {2022},
 }
@@ -49,7 +45,8 @@
   number={4},
   pages={827--841},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPDS.2018.2872064}
 }
 
 @misc{amdgpu_jl,
