### 2024-03-25 #137 @omlins
@mloubout :

> Staggered grid can lead to asymmetrical halos (potentially send only/recv only), a small explanation on how it is handled would be useful.

Thank you for pointing this out. In our approach, a field will only have halos in a given dimension if the corresponding overlap between the local fields is at least two cells wide; no halos are created if the overlap is only one cell wide (redundant computation is done instead), which avoids the need to deal with asymmetrical halos. We will see how to add this to the document.

>  How does it compare to similar DSLs for distributed stencil/pde solving such as Firedrake, Fenics, Devito

Thank you for your comment. We will be happy to do a comparison with related work as the one you mentioned, and with other key work in the vast fields of distributed memory parallelization and architecture-agnostic programming, as well as, more generally speaking, with important work that shares our objective to contribute to solving the challenge of the 3 “P”s -- (scalable) Performance, (performance) Portability and Productivity --.  However, given the large amount of work our contribution is related to, we aim to do so in a full paper submission. We did not have the space to do so in this very short extended abstract format of "at most two pages of content *including references*" (see author's guide: https://juliacon.github.io/proceedings-guide/author/), which "lays out in a concise fashion the methodology and use cases of the work presented at the conference", as opposed to a full paper submission, which "compared to an extended abstract, (...) compares the work to other approaches taken in the field and gives some additional insights on the conference contribution".
