+++ paper/ref.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @INPROCEEDINGS{wse_stencil,
@@ -49,7 +50,8 @@
   number={4},
   pages={827--841},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPDS.2018.2872064},
 }
 
 @misc{amdgpu_jl,
