+++ paper/ref.bib
@@ -23,10 +23,9 @@
 
 @article{parallelstencil2022,
   author={Omlin, S. and R{\"{a}}ss, L.},
-  doi = {???},
+  doi = {10.48550/arXiv.2211.15634},
   journal = {Proc. JuliaCon Conf.},
   title = {{H}igh-performance x{PU} {S}tencil {C}omputations in {J}ulia},
-  volume = {?},
   pages = {2},
   year = {2022},
 }
