+++ paper/ref.bib
@@ -57,7 +57,8 @@
   title={{AMDGPU.jl}: {AMD} {GPU} ({ROC}m) programming in {J}ulia },
   author={Samaroo, J. and Besard, T. and Churavy, V. and Lin, D. and other contributors},
   year={2013},
-  howpublished={\url{https://github.com/JuliaGPU/AMDGPU.jl}}
+  howpublished={\url{https://github.com/JuliaGPU/AMDGPU.jl}},
+  doi={10.5281/zenodo.10040461}
 }
 
 @online{juliacon2020scaling,
