### 2024-10-18 #137 @georgebisbas
Hi!
Thanks again for the improvements. A few more points that could be addressed:

- The caption of Figure 2 could also be enhanced with more details on the executed problem (similar to Figure 3)

> We do not have any use cases where we have a strictly fixed problem size. Thus, we considered presenting weak scaling more important.

I highly appreciate that, but why not presenting a 1024^3 problem of heat diffusion, just up to the point where the local decomposed problem becomes small enough to hit a strong scaling plateau?
Weak scaling benchmarking often hides communication efficiency, since the ratio of computation to communication remains high and constant, similar to the smallest possible decomposition.
Strong scaling could help us see whether e.g. using 8 GPUs starts having a plateau for a 1024^3 problem.

> The single-node performance is very high thanks to the implementation using ParallelStencil.jl.

It would be good to add a pointer to the ParallelStencil.jl paper in the manuscript

- Community guidelines should be also added. Please correct if I have not seen them so far.
https://github.com/JuliaCon/proceedings-review/issues/167
