### 2024-10-28 279ec7e
extend caption of figure 2 and add pointer to PS 2024
