+++ paper/ref.bib
@@ -21,13 +21,17 @@
   doi={10.1109/SC41405.2020.00062}
 }
 
-@article{parallelstencil2022,
+@article{parallelstencil2024,
+  doi = {10.21105/jcon.00138},
+  url = {https://doi.org/10.21105/jcon.00138},
+  year = {2024},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {64},
+  pages = {138},
   author={Omlin, S. and R{\"{a}}ss, L.},
-  doi = {10.48550/arXiv.2211.15634},
-  journal = {Proc. JuliaCon Conf.},
   title = {{H}igh-performance x{PU} {S}tencil {C}omputations in {J}ulia},
-  pages = {2},
-  year = {2022},
+  journal = {Proceedings of the JuliaCon Conferences}
 }
 
 @inproceedings{byrne2021mpi,
