### 2024-10-28 #137 @omlins
@georgebisbas: Thank you very much for the additional comments. In the following you can find the replies:

> The caption of Figure 2 could also be enhanced with more details on the executed problem (similar to Figure 3)

We have extended the caption of Figure 2 with more information similar to Figure 3.

--
> > The single-node performance is very high thanks to the implementation using ParallelStencil.jl.

> It would be good to add a pointer to the ParallelStencil.jl paper in the manuscript

We have added the pointer to the ParallelStencil.jl paper in the caption of Figure 2.

--
> Community guidelines should be also added. Please correct if I have not seen them so far.
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support #167

We have added guidelines for contribution to the readme.

--
> > > Minor comments (not aiming to be addressed, but possibly interesting points for future work and further discussion):
> > > - Why is strong scaling not presented in this work?

> > We do not have any use cases where we have a strictly fixed problem size. Thus, we considered presenting weak scaling more important.

> I highly appreciate that, but why not presenting a 1024^3 problem of heat diffusion, just up to the point where the local decomposed problem becomes small enough to hit a strong scaling plateau?

Thank you for your suggestion. It could certainly be interesting to dive into a strong scaling analysis. However, we believe that this does not fit into this very short extended abstract of two pages. Thus, we would like to follow your original suggestion and leave this for future work.
