### 2024-10-31 #137 @luraess
@fcdimitr please wait until I update the meta files and info to solve the non-showing DOI and special character issue before taking further action. 🙏
