### 2024-10-31 1aa6da3
Update DOI display and fix author name
