+++ paper/header.tex
@@ -3,7 +3,7 @@
 \title{Distributed Parallelization of xPU Stencil Computations in Julia}
 
 \author[1]{Samuel Omlin}
-\author[2, 3]{Ludovic R\"ass}
+\author[2, 3]{Ludovic Räss}
 \author[2, 3]{Ivan Utkin}
 \affil[1]{Swiss National Supercomputing Centre (CSCS), ETH Zurich, Lugano, Switzerland}
 \affil[2]{Laboratory of Hydraulics, Hydrology and Glaciology (VAW), ETH Zurich, Zurich, Switzerland}
@@ -13,8 +13,8 @@
 
 \hypersetup{
 pdftitle = {Distributed Parallelization of xPU Stencil Computations in Julia},
-pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {Samuel Omlin, Ludovic R\"ass, Ivan Utkin},
+pdfsubject = {JuliaCon 2022 Proceedings},
+pdfauthor = {Samuel Omlin, Ludovic Räss, Ivan Utkin},
 pdfkeywords = {Julia, Distributed Parallelization, xPU, GPU, Supercomputing, Stencil Computations, Staggered Grid},
 }
 
