+++ paper/paper.yml
@@ -11,7 +11,7 @@ authors:
   - name: Samuel Omlin
     orcid: 0000-0001-8969-5619
     affiliation: 1
-  - name: Ludovic R\"ass
+  - name: Ludovic Räss
     orcid: 0000-0002-1136-899X
     affiliation: 2, 3
   - name: Ivan Utkin
