### 2024-11-04 #137 @luraess
@fcdimitr The DOI and author name related issues are solved. Thanks for waiting. Since there seems to be no remaining issues, we could proceed.
