### 2024-11-05 #137 @fcdimitr
Thank you @luraess! What is the Zeonodo archival DOI?

:wave: @mloubout and @georgebisbas. Could you please mark all checkboxes in your list as done? Thank you!
