### 2024-11-05 #137 @georgebisbas
> Thank you @luraess! What is the Zeonodo archival DOI?
>
> 👋 @mloubout and @georgebisbas. Could you please mark all checkboxes in your list as done? Thank you!

Done, thanks
