### 2024-11-08 #137 @omlins
@fcdimitr : here you can find the DOI for the paper: https://doi.org/10.5281/zenodo.14056962
here you can find the DOI for the code: https://doi.org/10.5281/zenodo.13847550
